+++ Cryptography/inverse.c
@@ -18,23 +18,23 @@ void scan_for_inverse(int *a, int *n) {
 void print_inverse(int a, int n) {
 	int g = gcd(a, n);
 	printf("%dx === %d (mod %d)\n", a, g, n);
-	int *inv = inverse(a, n);
+	long long *inv = inverse(a, n);
 	if (inv == NULL) {
 		printf("Can't find the inverse between %d and %d\n", a, n);
 		printf("because the gcd between them is %d\n", g);
 		printf("and it isn't equal to 1\n");
 		return;
 	}
-	printf("x === %d (mod %d)\n", *inv, n);
+	printf("x === %lld (mod %d)\n", *inv, n);
 	free(inv);
 }
 
-int *inverse(int a, int n) {
-	int gcd, x, y;
+long long *inverse(long long a, long long n) {
+	long long gcd, x, y;
 	extended_gcd(a, n, &gcd, &x, &y);
 	if (gcd != 1) return NULL;
 	a %= n;
-	int *toRet = malloc(sizeof(int));
+	long long *toRet = malloc(sizeof(long long));
 	*toRet = (x < 0) ? x + n : x;
 	return toRet;
 }
